martello i toth - po BigDecimal, dziwny L2

+++ src/common/Constants.java
@@ -1,8 +1,10 @@
 package common;
 
+import java.math.BigDecimal;
+
 /**
  * Created by michto on 24.10.15.
  */
 public class Constants {
-    public static final double MAX_CAPACITY_OF_BOX = 1;
+    public static final BigDecimal MAX_CAPACITY_OF_BOX = BigDecimal.valueOf(1);
 }
